tsundokuEntityにcategoryId追加&ConfirmScreenの整理
