+++ app/src/main/java/com/example/tsundokun/data/local/dao/TsundokuDao.kt
@@ -14,12 +14,12 @@ interface TsundokuDao {
     fun observeById(id: String): Flow<TsundokuEntity>
 
     // カテゴリーから積読の取得(リアルタイム)
-    @Query("SELECT tsundoku.id, tsundoku.link, tsundoku.is_read, tsundoku.is_favorite, tsundoku.created_at, tsundoku.update_at, tsundoku.deleted_at FROM tsundoku INNER JOIN tsundoku_category ON tsundoku.id = tsundoku_category.tsundoku_id WHERE tsundoku_category.category_id = :category_id")
-    fun observeTsundokuByCategoryId(category_id: String): Flow<List<TsundokuEntity>>
+//    @Query("SELECT tsundoku.id, tsundoku.link, tsundoku.is_read, tsundoku.is_favorite, tsundoku.created_at, tsundoku.update_at, tsundoku.deleted_at FROM tsundoku INNER JOIN tsundoku_category ON tsundoku.id = tsundoku_category.tsundoku_id WHERE tsundoku_category.category_id = :category_id")
+//    fun observeTsundokuByCategoryId(category_id: String): Flow<List<TsundokuEntity>>
 
     // カテゴリーから積読の取得
-    @Query("SELECT tsundoku.id, tsundoku.link, tsundoku.is_read, tsundoku.is_favorite, tsundoku.created_at, tsundoku.update_at, tsundoku.deleted_at FROM tsundoku INNER JOIN tsundoku_category ON tsundoku.id = tsundoku_category.tsundoku_id WHERE tsundoku_category.category_id = :category_id")
-    fun getTsundokuByCategoryId(category_id: String): List<TsundokuEntity>
+//    @Query("SELECT tsundoku.id, tsundoku.link, tsundoku.is_read, tsundoku.is_favorite, tsundoku.created_at, tsundoku.update_at, tsundoku.deleted_at FROM tsundoku INNER JOIN tsundoku_category ON tsundoku.id = tsundoku_category.tsundoku_id WHERE tsundoku_category.category_id = :category_id")
+//    fun getTsundokuByCategoryId(category_id: String): List<TsundokuEntity>
 
     @Query("SELECT * FROM tsundoku")
     suspend fun getAll(): List<TsundokuEntity>
