+++ app/src/main/java/com/example/tsundokun/domain/models/Tsundoku.kt
@@ -11,4 +11,5 @@ data class Tsundoku(
     val isRead: Boolean,
     val updatedAt: String,
     val deletedAt: String,
+    val categoryId: String,
 )
