+++ app/src/main/java/com/example/tsundokun/domain/usecases/GetTsundokuUseCase.kt
@@ -1,5 +1,6 @@
 package com.example.tsundokun.domain.usecases
 
+import androidx.compose.ui.platform.isDebugInspectorInfoEnabled
 import com.example.tsundokun.data.local.entities.CategoryEntity
 import com.example.tsundokun.data.local.repository.TsundokuRepository
 import javax.inject.Inject
@@ -9,7 +10,4 @@ class GetTsundokuUseCase @Inject constructor(
 ) {
     val observeAllTsundoku = tsundokuRepository.observeAllTsundoku()
     val observeAllCategory = tsundokuRepository.observeAllCategory()
-
-    //addCategoryの呼び出し
-
 }
