+++ app/src/main/java/com/example/tsundokun/ui/confirm/component/data/ConfirmScreenNavArgs.kt
@@ -1,0 +1,7 @@
+package com.example.tsundokun.ui.confirm.component.data
+
+data class ConfirmScreenNavArgs(
+    val link: String,
+    var createdAt: String,
+    var categoryId: String,
+)
