+++ app/src/main/java/com/example/tsundokun/ui/home/category/tab/CategoryTab.kt
@@ -57,7 +57,7 @@ fun CategoryTab(navigator: DestinationsNavigator, categories: List<CategoryEntit
             TsundokuListScreen(selectedCategoryTsundokuList = webPageList.filter { it.isFavorite }, navigator = navigator)
         }
         else -> {
-            TsundokuListScreen(selectedCategoryTsundokuList = webPageList.filter { it.id.equals(tabSelected) }, navigator = navigator)
+            TsundokuListScreen(selectedCategoryTsundokuList = webPageList.filter { it.categoryId.equals(tabSelected) }, navigator = navigator)
         }
     }
     TsundokuListScreen(selectedCategoryTsundokuList = webPageList, navigator = navigator)
